requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
kombu==5.2.4
mccabe==0.7.0
prompt-toolkit==3.0.33
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
